* develop: (28 commits)
  EHUFA-NONE : Fixing the merge issue.
  EHUFA-1247 : Added signin config for debug to avoid uninstall when you run tests.
  EHUFA-1247 : Merged develop to feature branch.
  EHUFA-1247 : Merged develop to feature branch.
  EHUFA-1247 :  Fixing review comments for viewpropertymatcher and fixing affected tests.
  EHUFA-1247 :  Fixing review comments
  EHUFA-1247 :  Fixing review comments
  EHUFA-1247 : Added colors for navigation bar as per themeing.
  EHUFA-1247 :  Menu item text color is as per navigation bar text color known issue below kitkat https://code.google.com/p/android/issues/detail?id=172591&thanks=172591&ts=1431349509
  EHUFA-1247 : Formatting and renamed attribute.
  EHUFA-1247 : Corrected typo in styles.
  EHUFA-1247 :  Added tests.
  EHUFA-1247 : Added work around for title when custom title needs to be used.
  EHUFA-1247 : Added style for menu text appearance
  EHUFA-1247 : Added automated tests for title bar and menu icon sizes.
  EHUFA-1247 : Added automated tests for title bar and menu icon sizes.
  EHUFA-1247 : Navigation bar implementation as per android behavior.
  EHUFA-1247 : renamed packages in test folder, added few tests for toolbar.
  EHUFA-1247 : Added specification as per design for action buttons.
  EHUFA-1247 : Removed redundant code.
  ...

# Conflicts:
#	Source/CatalogApp/app/src/main/res/layout/preview_toolbar.xml
#	Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/components/Alert/AlertTest.java
#	Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/components/Buttons/PrimaryButtonWithTextOnlyTest.java
#	Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/components/Buttons/QuietButtonTest.java
#	Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/components/textbox/TextBoxTest.java
#	Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/drawableutils/BaseStateColorsImpl.java
#	Source/UIKit/uit/src/main/res/values/attrs.xml
#	Source/UIKit/uit/src/main/res/values/colors.xml
#	Source/UIKit/uit/src/main/res/values/dimens.xml
#	Source/UIKit/uit/src/main/res/values/styles.xml
#	Source/UIKit/uit/src/main/res/values/uid_alert_dialog_style.xml
#	Source/UIKit/uit/src/main/res/values/uit_actionbar_styles.xml
#	Source/UIKit/uit/src/main/res/values/uit_theme_interface_attrs.xml
#	Source/UIKit/uit/src/main/res/values/uit_theme_variant.xml

+++ Source/UIKit/uit/src/androidTest/java/com/philips/platform/uit/components/Alert/AlertTest.java
@@ -53,7 +53,7 @@ public class AlertTest {
      *********************************************/
     @Test
     public void verifyAlertWidth() {
-        int expectedWidth = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_width));
+        int expectedWidth = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_width);
         getAlert().check(matches(ViewPropertiesMatchers.isSameViewWidth(expectedWidth)));
     }
 
@@ -63,46 +63,46 @@ public class AlertTest {
 
     @Test
     public void verifyAlertTitleFontSize() {
-        int expectedFontSize = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_title_font_size));
+        int expectedFontSize = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_title_font_size);
         getAlertTitle().check(matches(TextViewPropertiesMatchers.isSameFontSize(expectedFontSize)));
     }
 
     @Test
     public void verifyAlertTitleIconHeight() {
         waitFor(testResources, 750);
-        int expectedIconHeight = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_iconsize);
+        int expectedIconHeight = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_iconsize);
         getAlertTitleIcon().check(matches(ViewPropertiesMatchers.isSameViewMinHeight(expectedIconHeight)));
     }
 
     @Test
     public void verifyAlertTitleIconWidth() {
         waitFor(testResources, 750);
-        int expectedIconWidth = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_iconsize);
+        int expectedIconWidth = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_iconsize);
         getAlertTitleIcon().check(matches(ViewPropertiesMatchers.isSameViewMinWidth(expectedIconWidth)));
     }
 
     @Test
     public void verifyAlertTitleIconRightPadding() {
         waitFor(testResources, 750);
-        int expectedRightPadding = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_icon_rightpadding));
+        int expectedRightPadding = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_icon_rightpadding);
         getAlertTitleIcon().check(matches((ViewPropertiesMatchers.isSameRightPadding(expectedRightPadding))));
     }
 
     @Test
     public void verifyAlertHeaderTopMargin() {
-        int expectedTopMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_icon_leftrighttop_margin);
+        int expectedTopMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_icon_leftrighttop_margin);
         getAlertTitleIcon().check(matches(ViewPropertiesMatchers.isSameTopMargin(expectedTopMargin)));
     }
 
     @Test
     public void verifyAlertHeaderRightMargin() {
-        int expectedRightMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_leftrighttop_margin);
+        int expectedRightMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_leftrighttop_margin);
         getAlertHeader().check(matches(ViewPropertiesMatchers.isSameRightMargin(expectedRightMargin)));
     }
 
     @Test
     public void verifyAlertHeaderLeftMargin() {
-        int expectedLeftMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_leftrighttop_margin);
+        int expectedLeftMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitle_leftrighttop_margin);
         getAlertHeader().check(matches(ViewPropertiesMatchers.isSameLeftMargin(expectedLeftMargin)));
     }
 
@@ -112,25 +112,25 @@ public class AlertTest {
 
     @Test
     public void verifyAlertContentFontSize() {
-        int expectedFontSize = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_font_size));
+        int expectedFontSize = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_font_size);
         getAlertContent().check(matches(TextViewPropertiesMatchers.isSameFontSize(expectedFontSize)));
     }
 
     @Test
     public void verifyAlertContentTextLeading() {
-        int expectedTextLeading = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontenttext_linespacing));
+        int expectedTextLeading = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontenttext_linespacing);
         getAlertContent().check(matches(TextViewPropertiesMatchers.isSameLineSpacing(expectedTextLeading)));
     }
 
     @Test
     public void verifyAlertTitleTextLeading() {
-        int expectedTextLeading = (int) (testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitletext_linespacing));
+        int expectedTextLeading = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alerttitletext_linespacing);
         getAlertTitle().check(matches(TextViewPropertiesMatchers.isSameLineSpacing(expectedTextLeading)));
     }
 
     @Test
     public void verifyAlertContentTopMargin() {
-        int expectedTopMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_top_padding);
+        int expectedTopMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_top_padding);
         getAlertHeader().check(matches(ViewPropertiesMatchers.isSameBottomMargin(expectedTopMargin)));
     }
 
@@ -143,13 +143,13 @@ public class AlertTest {
 
     @Test
     public void verifyAlertContentRightMargin() {
-        int expectedRightMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_leftrightbottom_margin);
+        int expectedRightMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_leftrightbottom_margin);
         getAlertContent().check(matches(ViewPropertiesMatchers.isSameRightMargin(expectedRightMargin)));
     }
 
     @Test
     public void verifyAlertContentBottomMargin() {
-        int expectedBottomMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_leftrightbottom_margin);
+        int expectedBottomMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertcontent_leftrightbottom_margin);
         getAlertActionArea().check(matches(ViewPropertiesMatchers.isSameTopMargin(expectedBottomMargin)));
     }
 
@@ -167,7 +167,7 @@ public class AlertTest {
         mActivityTestRule.getActivity().switchFragment(DialogTestFragment.create());
         getAlertTitle().check(matches(ViewPropertiesMatchers.isVisible(View.GONE)));
 
-        int expectedTopMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_content_top_margin_when_no_title);
+        int expectedTopMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alert_content_top_margin_when_no_title);
         getAlertContainer().check(matches(ViewPropertiesMatchers.isSameTopMargin(expectedTopMargin)));
     }
 
@@ -177,13 +177,13 @@ public class AlertTest {
 
     @Test
     public void verifyPaddingBetweenActionButtons() {
-        int expectedButtonsMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertaction_buttons_padding);
+        int expectedButtonsMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertaction_buttons_padding);
         getAlertConfirmativeButton().check(matches(ViewPropertiesMatchers.isSameLeftMargin(expectedButtonsMargin)));
     }
 
     @Test
     public void verifyRightPaddingOfActionButtonView() {
-        int expectedButtonRightMargin = (int) testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertaction_button_rightpadding);
+        int expectedButtonRightMargin = testResources.getDimensionPixelSize(com.philips.platform.uit.test.R.dimen.alertaction_button_rightpadding);
         getAlertConfirmativeButton().check(matches(ViewPropertiesMatchers.isSameRightMargin(expectedButtonRightMargin)));
     }
